refactor(compiler): allow different kinds of directive matchers to be passed to binder
Updates the target binder to allow either a selector-based or selectorless matcher to be passed in. This will allow us to skip some of the overhead when matching directives to nodes.
